refactor: Bridge에 사용되는 타입들을 type에 따로 뺴서 정의
#10
