+++ src/components/common/entities/Bridge/index.tsx
@@ -1,31 +1,10 @@
 "use client";
 
+import {
+  MessageEventRequestData,
+  MessageEventResponseData,
+} from "@/type/bridge";
 import { useEffect } from "react";
-
-declare global {
-  interface Window {
-    ReactNativeWebView?: {
-      postMessage(message: string): void;
-    };
-  }
-  interface Document {
-    ReactNativeWebView?: {
-      postMessage(message: string): void;
-    };
-  }
-}
-
-export interface MessageEventResponseData<Data = unknown> {
-  status: "success" | "error";
-  name: string;
-  data?: Data;
-}
-
-export interface MessageEventRequestData<Body = unknown> {
-  method: "GET" | "POST" | "PUT" | "DELETE";
-  name: string;
-  body?: Body;
-}
 
 interface BridgeProps {
   onRequest: (reqMessage: MessageEventRequestData) => MessageEventResponseData;
